+++ CHANGELOG.md
@@ -1,12 +1,16 @@
+<!-- markdownlint-disable MD012 MD013 MD024 MD033 -->
 # Change Log
 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](https://semver.org/)
 
-## [Unreleased]
+## [0.5.0] 2025-11-03
 
 ### Added
 
-- initial version
+- initial version of the Search Logs task
+  - Service URL, Account and token parameter for authn
+  - Query, Time Range, Limit and Repositories parameter for limiting search scope and result
+  - List of output paths paramter to specify the output schema